Split markdown block cache type and constant into their own files
