+++ packages/twenty-front/src/modules/ai/types/MarkdownBlockSplitCache.ts
@@ -1,0 +1,6 @@
+export type MarkdownBlockSplitCache = {
+  text: string;
+  blocks: string[];
+  stablePrefix: string;
+  stableBlocks: string[];
+};
